feat: create separate directory for handlers
